requirements.txt
# Core ML & Data Science
dvc[azure]
pandas
numpy
scikit-learn
scipy
joblib
evidently

# ML Frameworks
xgboost
torch
torchvision
imbalanced-learn

# API & Web
fastapi[standard]
uvicorn[standard]
python-multipart
azure-storage-blob
fastapi-users[sqlalchemy]
PyYAML
azure-storage-blob
opencensus-ext-azure
opencensus-ext-logging


# Database
psycopg2-binary
sqlalchemy
alembic

# ML Tracking & Experimentation
mlflow
optuna
optuna-integration[mlflow]

# Configuration & Validation
pyyaml
python-dotenv
pydantic
great-expectations

# HTTP & Networking
requests
httpx

# Development & Testing
pytest
pytest-asyncio
requests-mock
pytest-cov
setuptools

azure-ai-ml 
azure-identity 
azure-mgmt-resource 
azure-mgmt-machinelearningservices